Setting Up CI-CD

+++ Sources/TaskBoard.Application/UserCases/Boards/DeleteBoardHandler.cs
@@ -1,4 +1,5 @@
-﻿using TaskBoard.Domain.Exceptions;
+﻿using Serilog;
+using TaskBoard.Domain.Exceptions;
 using TaskBoard.Domain.Interfaces;
 
 namespace TaskBoard.Application.UseCases.Boards;
@@ -12,13 +13,23 @@ public class DeleteBoardHandler
         _boardRepository = boardRepository;
     }
 
-    public async Task Handle(Guid boardId)
+    public async Task Handle(Guid boardId, Guid userId)
     {
-        var board = await _boardRepository.GetByIdAsync(boardId);
+        var board = await _boardRepository.GetByIdAsync(boardId, userId);
 
         if (board is null)
             throw new NotFoundException($"Board with ID {boardId} not found.");
 
+        if (board.UserId != userId)
+            throw new UnauthorizedAccessException("You do not own this board.");
+
         await _boardRepository.DeleteAsync(board);
+
+        Log.Information("BoardDeleted {@Board}", new
+        {
+            board.Id,
+            board.UserId,
+            board.Name
+        });
     }
 }
